move debug to utils
