+++ packages/nx-firebase/src/utils/debug.ts
